\name{convertPicture}
\alias{convertPicture}
\title{
  Convert a vector image into Cairo PostScript or SVG.
}
\description{

  In order to read in an image, grImport and grImport2 first require
  that the images they read in are their target input formats. In the
  case of grImport it is simply a PostScript image, while grImport2
  requires an SVG image that has been generated by the Cairo graphics
  library. This function provides an interface for performing the
  conversion from a source vector image format to Cairo-generated
  PostScript or SVG.

}
\usage{
convertPicture(file, outfile,
               inputFormat = NA, outputFormat = NA,
               pagenum = 1L)
}
\arguments{
  \item{file}{
    A character vector representing the path to the input file.
  }
  \item{outfile}{
    A character vector representing the path to the output file.
  }
  \item{inputFormat}{
    A character vector representing the input image format. Must be one
    of \code{"ps"}, \code{"eps"}, \code{"pdf"} or \code{"svg"}. If this
    parameter is \code{NA} then an attempt will be made to determine the
    image format automatically.
  }
  \item{outputFormat}{
    A character vector representing the output image format. Must be one
    of \code{"ps"}, \code{"eps"}, or \code{"svg"}. If this parameter is
    \code{NA} then an attempt will be made to determine the image format
    automatically.
  }
  \item{pagenum}{
    When using an input format with multiple pages present
    (e.g. PostScript or PDF), then this parameter determines which page
    will be converted to Cairo SVG.
  }
}
\details{
  When installing this package, it is not strictly necessary to have all
  of the image renderers available. In the case that only some of the
  renderers are installed, the function may provide an error message
  when the appropriate renderer is not available. For example, this may
  occur if it is not possible to render SVG images and yet the input
  file to this function is an SVG image.
}
\value{
  None. The side effect is that a new PostScript or SVG image is created.
}
\author{
  Simon Potter
}
